Add files via upload
Hey Folks! The above project is a password manager and is built using pure python and JSON. It has many features like copying the password to clipboard when generated or added, saving your passwords and user-id in a text field, and many more. I hope you like it and I request you to contribute to the project and make it helpful for other developers. THANK YOU!!